User login redirection and filter modal
