+++ next.config.ts
@@ -15,15 +15,15 @@ const nextConfig: NextConfig = {
   typescript: {
     ignoreBuildErrors: true,
   },
-  // async redirects() {
-  //   return [
-  //     {
-  //       source: "/",
-  //       destination: "/login",
-  //       permanent: false, 
-  //     },
-  //   ];
-  // },
+  async redirects() {
+    return [
+      {
+        source: "/",
+        destination: "/login",
+        permanent: false, 
+      },
+    ];
+  },
 };
 
 export default nextConfig;
