+++ src/app/(pages)/checkout/[id]/page.tsx
@@ -24,7 +24,9 @@ const page = () => {
     const getSubscriptions = () => {
         setLoading(true)
         sendRequest({ url: `/plans/${id}/`, method: "GET" }).then(res => {
-            setPlan(res.data)
+            setPlan(res)
+            setLoading(false)
+        }).finally(() => {
             setLoading(false)
         })
     }
