+++ src/components/common/Icon.tsx
@@ -99,13 +99,13 @@ const Icon: React.FC<PropsTypes> = ({ name, fill, className, stroke, size = "2em
         ),
         proPlan: (
             <svg width="44" height="44" viewBox="0 0 44 44" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <path d="M5.5 14.6666L8.14537 30.5389C8.73472 34.0748 11.7941 36.6666 15.3789 36.6666H28.6211C32.2058 36.6666 35.2653 34.0748 35.8547 30.5389L38.5 14.6666M5.5 14.6666L12.386 20.3006C14.0868 21.6921 16.6214 21.2853 17.8013 19.4312L22 12.8333M5.5 14.6666C7.01879 14.6666 8.25 13.4354 8.25 11.9166C8.25 10.3978 7.01879 9.16659 5.5 9.16659C3.98121 9.16659 2.75 10.3978 2.75 11.9166C2.75 13.4354 3.98121 14.6666 5.5 14.6666ZM38.5 14.6666L31.614 20.3006C29.9132 21.6921 27.3786 21.2853 26.1987 19.4312L22 12.8333M38.5 14.6666C40.0187 14.6666 41.25 13.4354 41.25 11.9166C41.25 10.3978 40.0187 9.16659 38.5 9.16659C36.9813 9.16659 35.75 10.3978 35.75 11.9166C35.75 13.4354 36.9813 14.6666 38.5 14.6666ZM22 12.8333C23.5187 12.8333 24.75 11.602 24.75 10.0833C24.75 8.56446 23.5187 7.33325 22 7.33325C20.4813 7.33325 19.25 8.56446 19.25 10.0833C19.25 11.602 20.4813 12.8333 22 12.8333Z" stroke="#003BFF" strokeWidth="2.5" stroke-linecap="round" stroke-linejoin="round" />
+                <path d="M5.5 14.6666L8.14537 30.5389C8.73472 34.0748 11.7941 36.6666 15.3789 36.6666H28.6211C32.2058 36.6666 35.2653 34.0748 35.8547 30.5389L38.5 14.6666M5.5 14.6666L12.386 20.3006C14.0868 21.6921 16.6214 21.2853 17.8013 19.4312L22 12.8333M5.5 14.6666C7.01879 14.6666 8.25 13.4354 8.25 11.9166C8.25 10.3978 7.01879 9.16659 5.5 9.16659C3.98121 9.16659 2.75 10.3978 2.75 11.9166C2.75 13.4354 3.98121 14.6666 5.5 14.6666ZM38.5 14.6666L31.614 20.3006C29.9132 21.6921 27.3786 21.2853 26.1987 19.4312L22 12.8333M38.5 14.6666C40.0187 14.6666 41.25 13.4354 41.25 11.9166C41.25 10.3978 40.0187 9.16659 38.5 9.16659C36.9813 9.16659 35.75 10.3978 35.75 11.9166C35.75 13.4354 36.9813 14.6666 38.5 14.6666ZM22 12.8333C23.5187 12.8333 24.75 11.602 24.75 10.0833C24.75 8.56446 23.5187 7.33325 22 7.33325C20.4813 7.33325 19.25 8.56446 19.25 10.0833C19.25 11.602 20.4813 12.8333 22 12.8333Z" stroke="#003BFF" strokeWidth="2.5" strokeLinejoin="round" strokeLinejoin="round" />
             </svg>
 
         ),
         user: (
             <svg width="20" height="17" viewBox="0 0 20 17" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <path d="M15.6666 16L13.0716 13.5M13.0716 13.5L15.6666 11M13.0716 13.5H18.2615M9.6117 11.4167H5.7193C4.51217 11.4167 3.9086 11.4167 3.41747 11.5602C2.31167 11.8834 1.44634 12.717 1.1109 13.7824C0.961914 14.2555 0.961914 14.837 0.961914 16M11.7742 4.75C11.7742 6.82107 10.0315 8.5 7.88175 8.5C5.73203 8.5 3.98934 6.82107 3.98934 4.75C3.98934 2.67893 5.73203 1 7.88175 1C10.0315 1 11.7742 2.67893 11.7742 4.75Z" stroke="#ACACAC" strokeWidth="1.5" stroke-linecap="round" stroke-linejoin="round" />
+                <path d="M15.6666 16L13.0716 13.5M13.0716 13.5L15.6666 11M13.0716 13.5H18.2615M9.6117 11.4167H5.7193C4.51217 11.4167 3.9086 11.4167 3.41747 11.5602C2.31167 11.8834 1.44634 12.717 1.1109 13.7824C0.961914 14.2555 0.961914 14.837 0.961914 16M11.7742 4.75C11.7742 6.82107 10.0315 8.5 7.88175 8.5C5.73203 8.5 3.98934 6.82107 3.98934 4.75C3.98934 2.67893 5.73203 1 7.88175 1C10.0315 1 11.7742 2.67893 11.7742 4.75Z" stroke="#ACACAC" strokeWidth="1.5" strokeLinejoin="round" strokeLinejoin="round" />
             </svg>
 
         ),
